ZEPPELIN-196: fix the auto-completion issue
The existing code  ($rootScope.$emit('sendNewEvent',...)) emits events to the  "sendNewEvent" channel but apparently there are no listener listening to it!

The websocketMsgSrv service exposes a completion() method that is not used. The WS communication goes through this webscket service so it seemed natural to use it for the completion method as well.

This change fixes the event flow problem. You can trace that COMPLETION events go down to the Interceptor's compete method and back to the completers $scope.$on('completionList') handler.

It seems though that the SparkInteprteer's completer doesn't work. It never returns completion suggestions. But this is would be a different issue.

Author: tzolov <christian.tzolov@gmail.com>

Closes #180 from tzolov/ZEPPELIN-196 and squashes the following commits:

e73223f [tzolov] ZEPPELIN-196: fix the autocompletion issue

(cherry picked from commit 9186b2186de157e2ac4d9ff13ebfc23c979b9c51)
Signed-off-by: Lee moon soo <moon@apache.org>